// src/app/Models/GetDeptHumanDataClass.ts
// Generated by https://quicktype.io

export class GetDeptHumanDataClass {
    DeptHumanID: number;
    DeptCode:    string;
    DeptName:    string;
    EmpID:       string;
    EmpName:     string;
    IsValid:     boolean;
    Note:        string;
    UpdateMan:   string;
    UpdateDate:  string;
}