// src/app/Models/GetDeptDataClass.ts
// Generated by https://quicktype.io

export class GetDeptDataClass {
    DeptID:        number;
    DeptCode:      string;
    DeptNameC:     string;
    ParentID:      number;
    ParentName:    string;
    Tree:          string;
    DateA:         string;
    DateD:         string;
    Manage:        string;
    ManageMail:    string;
    Path:          string;
    Level:         number;
    Superut:       any;
    Unitlvl:       number;
    AbsLimit:      number;
    WeekStartDate: string;
    ChildDept:     any;
    Base:          any;
    uiShowDeptCodeAndName:string
}